+++ pixiu/tester/tester_api_v1.py
@@ -5,7 +5,7 @@ import talib
 import logging
 from datetime import datetime
 from pixiu.api.v1 import (TimeFrame, OrderCommand, OrderType, IndicatiorID, SymbolIndicator, SymbolPrice, SymbolData,
-                      Order, Account, Symbol, AccountData, ErrorID, OrderResult)
+                      Order, Account, Symbol, AccountData, ErrorID, OrderResult, OrderScope)
 log = logging.getLogger(__name__)
 
 
@@ -163,16 +163,16 @@ class TesterAPI_V1(API_V1_Base):
     def Symbol(self, shift=0) -> str:
         return self.tester.symbol
 
-    def GetOpenedOrderUIDs(self, symbol: str = None):
-        ret = self.tester.get_order_dict(symbol, status="opened", script_name=None)
+    def GetOpenedOrderUIDs(self, symbol: str = None, scope: int = OrderScope.EA):
+        ret = self.tester.get_order_dict(symbol, status="opened", scope=scope)
         return list(ret.keys())
 
-    def GetPendingOrderUIDs(self, symbol: str = None):
-        ret = self.tester.get_order_dict(symbol, status="pending", script_name=None)
+    def GetPendingOrderUIDs(self, symbol: str = None, scope: int = OrderScope.EA):
+        ret = self.tester.get_order_dict(symbol, status="pending", scope=scope)
         return list(ret.keys())
 
-    def GetClosedOrderUIDs(self, symbol: str = None):
-        ret = self.tester.get_order_dict(symbol, status="closed", script_name=None)
+    def GetClosedOrderUIDs(self, symbol: str = None, scope: int = OrderScope.EA):
+        ret = self.tester.get_order_dict(symbol, status="closed", scope=scope)
         return list(ret.keys())
 
     def StopTester(self, code: int=0, message: str=None):